Refactor some board model logic into board_model.py
